# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: DokuSuite API
  version: 0.2.3
  description: |
    API für iOS-App, Web-Tool und Integrationen der DokuSuite.
    Zeitzone: Europe/Berlin. Datums-/Zeitangaben sind RFC3339 UTC, mit klarer Interpretation für Woche (ISO-8601 Kalenderwoche).
servers:
  - url: /api
...
security:
  - bearerAuth: []
paths:
...
  /shares:
    get:
      tags: [shares]
      summary: List shares (read-only)
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - in: query
          name: orderId
          schema: { type: integer }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Page_Share_' } } } }
    post:
      tags: [shares]
      summary: Create a customer share for an order
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ShareCreate' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Share' } } } }
  /shares/{id}:
    get:
      tags: [shares]
      summary: Get share
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Share' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /shares/{id}/revoke:
    post:
      tags: [shares]
      summary: Revoke share
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '204': { description: Revoked }
        '404': { $ref: '#/components/responses/NotFound' }
  /public/shares/{token}/photos/{id}:
    get:
      tags: [public-shares]
      summary: Get public photo URLs
      security: []
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Pre-signed URLs
          content:
            application/json:
              schema:
                type: object
                properties:
                  original_url: { type: string, format: uri }
                  thumbnail_url: { type: string, format: uri }
        '404':
          $ref: '#/components/responses/NotFound'
  /public/shares/{token}/photos:
    get:
      tags: [public-shares]
      summary: List public photos
      security: []
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Photo ids
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicPhotoList' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      in: path
      name: id
      required: true
      schema: { type: integer }
    page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    limit:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
    PageMeta:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
...
    Page_Share_:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Share' }
        meta: { $ref: '#/components/schemas/PageMeta' }
...
    WatermarkPolicy:
      type: string
      enum: [none, default, custom_text]
...
    ShareCreate:
      type: object
      required: [order_id]
      properties:
        order_id: { type: integer }
        expires_at: { type: string, format: date-time, nullable: true }
        download_allowed: { type: boolean, default: true }
        watermark_policy: { $ref: '#/components/schemas/WatermarkPolicy', nullable: true }
        email: { type: string, format: email, nullable: true }
    Share:
      type: object
      properties:
        id: { type: integer }
        order_id: { type: integer }
        url: { type: string, format: uri }
        expires_at: { type: string, format: date-time, nullable: true }
        download_allowed: { type: boolean }
        watermark_policy: { $ref: '#/components/schemas/WatermarkPolicy', nullable: true }
    PublicPhoto:
      type: object
      properties:
        id: { type: integer }
    PublicPhotoList:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/PublicPhoto' }
...